Core logic revamp

+++ src/main/java/flat/animations/Animation.java
@@ -2,8 +2,10 @@ package flat.animations;
 
 public interface Animation {
 
+    // TODO - Animation source must be assigned to the same active that the animation si running
+
     boolean isPlaying();
 
-    void handle(long milis);
+    void handle(float milis);
 
 }
